pyproject.toml
[tool.poetry]
name = "passmanager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
